Updated makefile to different platforms
There are now 3 makefiles: two for intel icpc and mkl (32 and 64 bit) and
one for gcc + atlas. Copy the one you want to use to Makefile and
everything works as it should. Remember to change the Makefile.* and not
Makefile as only Makefile.* are in the git index. I'm working on a hook to
automatic update the Makefile.